Berger and Boos n=10  with conf=80%
Adjusting the FWER boundaries using Berger and Boos after finding the lower bound of the nuisance parameter for estimated correlations. The method is using the the sample size n=50 and confidence level 80%.

Adjusting the FWER boundaries using Berger and Boos after finding the lower bound of the nuisance parameter for estimated correlations. The method is using the the sample size n=50 and confidence level 80%.